+++ examples/src/main/java/org/bindgen/example/methods/MethodExample.java
@@ -46,6 +46,7 @@ public class MethodExample {
 		this.name = name;
 	}
 
+	@Override
 	public String toString() {
 		return "method";
 	}
@@ -62,7 +63,8 @@ public class MethodExample {
 		return false;
 	}
 
-	// Putting the @deprecated here ensures a warning would show up if this "to" prefix got recognized
+	// Putting the @deprecated here ensures a warning would show up if this "to"
+	// prefix got recognized
 	@Deprecated
 	public boolean tobacco() {
 		return false;
@@ -73,12 +75,14 @@ public class MethodExample {
 		return false;
 	}
 
-	// Returning a binding to myself should cause recursion or anything--used to cause errors
+	// Returning a binding to myself should cause recursion or anything--used to
+	// cause errors
 	public Binding<?> getBinding() {
 		return new MethodExampleBinding(this);
 	}
 
-	// Putting the @deprecated here ensures a warning would show up if this array was not skipped
+	// Putting the @deprecated here ensures a warning would show up if this array
+	// was not skipped
 	@Deprecated
 	public String[] getStrings() {
 		return null;