+++ examples/src/test/java/org/bindgen/example/GenericsExampleTest.java
@@ -1,6 +1,6 @@
 package org.bindgen.example;
 
-import junit.framework.Assert;
+import org.junit.Assert;
 import junit.framework.TestCase;
 
 public class GenericsExampleTest extends TestCase {