+++ processor/src/test/java/org/bindgen/processor/InnerClassTest.java
@@ -8,6 +8,7 @@ public class InnerClassTest extends AbstractBindgenTestCase {
 
 	@Test
 	public void testGenerateBindingsForInnerClasses() throws Exception {
+		this.setScope("org.bindgen,java.lang");
 		ClassLoader loader = this.compile("org/bindgen/processor/inner/SomeClass.java");
 
 		Class<?> actualClass = loader.loadClass("org.bindgen.processor.inner.SomeClass$InnerClass");
@@ -17,7 +18,10 @@ public class InnerClassTest extends AbstractBindgenTestCase {
 		assertNotNull(bindingClass);
 		assertMethodDeclared(bindingClass, "x");
 		assertMethodDeclared(bindingClass, "squared");
-		assertMethodNotDeclared(bindingClass, "foobar"); // does not have access to protected from different package (someClass)
+		assertMethodNotDeclared(bindingClass, "foobar"); // does not have access
+															// to protected from
+															// different package
+															// (someClass)
 	}
 
 	@Test
@@ -27,7 +31,8 @@ public class InnerClassTest extends AbstractBindgenTestCase {
 		Class<?> actualClass = loader.loadClass("org.bindgen.processor.inner.nastyClass$InnerClass");
 		assertNotNull(actualClass);
 
-		Class<?> bindingClass = loader.loadClass("org.bindgen.processor.inner.bindgen_nastyClass.InnerClassBindingPath");
+		Class<?> bindingClass = loader
+				.loadClass("org.bindgen.processor.inner.bindgen_nastyClass.InnerClassBindingPath");
 		assertNotNull(bindingClass);
 		assertMethodDeclared(bindingClass, "x");
 	}