+++ processor/src/test/template/org/bindgen/processor/circular/SelectedIndividual.java
@@ -3,12 +3,12 @@ package org.bindgen.processor.circular;
 import org.bindgen.Bindable;
 
 @Bindable
-public class SelectedIndividual
-{
-    private Individual<?> individual;
-    
-    public Organisation<?, ?> getOrganisation()
-    {
-        return null;
-    }
+public class SelectedIndividual {
+
+	@SuppressWarnings("unused")
+	private Individual<?> individual;
+
+	public Organisation<?, ?> getOrganisation() {
+		return null;
+	}
 }